change recall_window

+++ backend/app/app/schemas/fact.py
@@ -46,7 +46,7 @@ class SchedulerQuery(BaseModel):
     repetition_model: Repetition
     env: str
     user_id: int
-    target_window: int
+    target_window: float
 
 class UpdateRequestV2(BaseModel):
     user_id: int
